Upload from mac

+++ src/actions/ContactActions.js
@@ -1,34 +1,30 @@
-'use strict'
 import dispatcher from "../AppDispatcher";
 import  * as Api from "../utils/AppAPI";
 import axios from 'axios';
 import 'react-notifications/lib/notifications.css';
-import {NotificationContainer, NotificationManager} from 'react-notifications';
+import { NotificationManager } from 'react-notifications';
 const querystring = require('querystring');
 const AppConstants = require("../constants/AppConstants");
 const Messages = require("../constants/Messages");
 const LIMIT = 50;
 
 export function reloadItemsAsync() {
-  dispatcher.dispatch({type: AppConstants.CONTACT_FETCH});
   const url = Api.getContacts(0, LIMIT);
-  setTimeout(() => {
-    axios(url).then((data) => {
-      // dispatcher
-      dispatcher.dispatch({
-        type: AppConstants.RECEIVE_CONTACT,
-        items: data["data"]
-      });
+  axios(url).then((data) => {
+    // receive dispatcher
+    dispatcher.dispatch({
+      type: AppConstants.RECEIVE_CONTACT,
+      items: data["data"]
     });
-  }, 500);
+  });
 }
 
 export function create(item) {
   const url = Api.createContact();
   axios.post(url, querystring.stringify(item)).then((response) => {
-    // alert message
-    NotificationManager.success(Messages.SUCCESS_MESSAGE, Messages.SUCCESS_TITLE);
-    // dispatcher
+    // success message
+    NotificationManager.success(Messages.CREATE_SUCCESS_MESSAGE, Messages.SUCCESS_TITLE);
+    // create dispatcher
     dispatcher.dispatch({
         type: AppConstants.CONTACT_CREATE,
         item: response["data"],
@@ -39,15 +35,42 @@ export function create(item) {
   });
 }
 
+export function update(item) {
+  // item checking
+  if(item !== undefined && item != null) {
+    item.isCheck = true;
+    const url = Api.deleteContact(item._id);
+    console.log(url);
+    setTimeout(function(){
+      // success message
+      NotificationManager.success(Messages.UPDATE_SUCCESS_MESSAGE, Messages.SUCCESS_TITLE);
+      // update dispatcher
+      dispatcher.dispatch({
+          type: AppConstants.CONTACT_UPDATE,
+          item: item,
+      });
+    }, 500);
+  } else {
+    NotificationManager.error(Messages.ERROR_MESSAGE, Messages.ERROR_TITLE);
+  }
+}
+
 export function remove(item) {
-  item.isCheck = true;
-  const url = Api.deleteContact(item._id);
-  console.log(url);
-  // alert message
-  NotificationManager.success(Messages.SUCCESS_MESSAGE, Messages.SUCCESS_TITLE);
-  // dispatcher
-  dispatcher.dispatch({
-      type: AppConstants.CONTACT_DELETE,
-      item: item,
-  });
+  // item checking
+  if(item !== undefined && item != null) {
+    item.isCheck = true;
+    const url = Api.deleteContact(item._id);
+    console.log(url);
+    setTimeout(function(){
+      // success message
+      NotificationManager.success(Messages.DELETE_SUCCESS_MESSAGE, Messages.SUCCESS_TITLE);
+      // remove dispatcher
+      dispatcher.dispatch({
+          type: AppConstants.CONTACT_DELETE,
+          item: item,
+      });
+    }, 500);
+  } else {
+    NotificationManager.error(Messages.ERROR_MESSAGE, Messages.ERROR_TITLE);
+  }
 }
